requirements.txt
pika==1.1.0
psycopg2-binary